+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * [^] `RadioboxGroup` 组件重命名为 `RadioGroup`。**[Breaking change]**
 * [^] `Button` 组件的 `aux` 风格 `ui` 现在是默认样式，原有的默认样式更名为 `secondary`。**[Breaking change]**
 * [^] 为组件支持了 `:focus-ring` 的 polyfill，需要使用方自行引入。详见 [#121](https://github.com/ecomfe/veui/issues/121)**[Breaking change]**
+* [^] 优化了 `Uploader` 组件的部分 prop、slot 的命名。**[Breaking change]**
+* [-] 移除了 `BreadCrumb` 组件的 prop `routers`。**[Breaking change]**
 * [^] 修复 `RegionPicker` 浮层在特殊情况下显示上的问题。
 * [+] 为 `Table` 增加 `select-mode` prop，用于提供单选模式。
 * [^] 修复 `Dropdown` 在 `ui` 为 `link` 时的样式。